Update README.md
We define the TransferFCN_purelin function with inputs w, p, B to use purelin.
In the main program, we first read the input and output values of the iris data set and store them in two variables, p and t. We form the matrices w, b and define the maximum epoch and the evaluation and training performance vector.